+++ packages/rust-bindings/src/external_functions.rs
@@ -373,6 +373,18 @@ pub fn get_verifiable_credential_public_key_ext(
     .map_err(to_js_error)
 }
 
+#[wasm_bindgen(js_name = getVerifiableCredentialBackupEncryptionKey)]
+pub fn get_verifiable_credential_backup_encryption_key_ext(
+        seed_as_hex: HexString,
+        raw_net: &str,
+) -> JsResult<HexString> {
+    get_verifiable_credential_backup_encryption_key_aux(
+            seed_as_hex,
+            raw_net,
+        )
+        .map_err(to_js_error)
+}
+
 #[wasm_bindgen(js_name = serializeCredentialDeploymentPayload)]
 pub fn serialize_credential_deployment_payload_ext(
     signatures: &JsValue,